discord-bot/requirements.txt
aiohttp==3.12.15
beautifulsoup4==4.13.5
discord.py==2.6.3
playwright==1.55.0
python-dotenv==1.1.1